# Conflicts:
#	g3w-admin/base/__init__.py
#	g3w-admin/qdjango/api/projects/serializers.py

+++ g3w-admin/qdjango/vector.py
@@ -58,7 +58,7 @@ from qdjango.api.layers.filters import (
     FILTER_FID_PARAM
 )
 
-from .models import Layer, SessionTokenFilter, SessionTokenFilterLayer
+from .models import Layer, SessionTokenFilter, SessionTokenFilterLayer, FilterLayerSaved
 from .utils.data import QGIS_LAYER_TYPE_NO_GEOM
 from .utils.edittype import MAPPING_EDITTYPE_QGISEDITTYPE
 
@@ -389,6 +389,20 @@ class LayerVectorView(QGISLayerVectorViewMixin, BaseVectorApiView):
         else:
             request_data = request.query_params
 
+        def _get_sessiontokenfilter():
+            """
+            Get or create an instance of SessionTokeFilter model
+
+            :return: A tuple with the instance of SessioneTokeFilter model and a boolean state for created or retreive
+            :rtype: tuple
+            """
+            s, created = SessionTokenFilter.objects.get_or_create(
+                sessionid=sessionid,
+                defaults={'user': request.user if request.user.pk else None}
+            )
+
+            return s, created
+
         # parameters to check:
         # mode: create, update, delete
         # fidsin: fids list to filter
@@ -398,7 +412,48 @@ class LayerVectorView(QGISLayerVectorViewMixin, BaseVectorApiView):
         fidsin = request_data.get('fidsin')
         fidsout = request_data.get('fidsout')
 
+        token_data = {}
+
+
+        def _create_qgs_expr(s, fidsin=None, fidsout=None):
+            """
+            Create qgs expression to save in db
+
+            :param s: Instance of SessionTokenFilter model.
+            :param fidsin: String of features id to contain.
+            :param fidsout: String of features id to exlude.
+            :return: QGIS expression string
+            :rtype: str
+            """
+
+            expr = f'$id IN ({fidsin})' if fidsin else f'$id NOT IN ({fidsout})'
+            return f'{s.qgs_expr} AND {expr}' if s else expr
+
+        def _check_url_params_name_fid(mode):
+            """
+            Check for 'name' and 'fid' url parameters
+
+            :param mode: Value of 'mode' URL parameter
+            :return: Dict for model to save with fid or name key
+            :rtype: dict
+            """
+            name = request_data.get('name')
+            fid = request_data.get('fid')
+            if not name and not fid:
+                raise APIException(
+                    f"'name' or 'fid' parameter is required for mode='{mode}'.")
+            if fid and name:
+                raise APIException(
+                    f"'fid' only or 'name' only parameter is required for mode='{mode}'.")
+
+            # Return dict for model
+            return {'pk': fid} if fid else {'name': name}
+
+
+
+
         if mode == 'create_update':
+
             if not fidsin and not fidsout:
                 raise APIException(
                     "'fidsin' or 'fidsout' parameter is required.")
@@ -407,24 +462,7 @@ class LayerVectorView(QGISLayerVectorViewMixin, BaseVectorApiView):
                 raise APIException(
                     "'fidsin' only or 'fidsout' only parameter is required.")
 
-        s, created = SessionTokenFilter.objects.get_or_create(
-            sessionid=sessionid,
-            defaults={'user': request.user if request.user.pk else None}
-        )
-
-        token_data = {}
-
-        if mode == 'delete' and not s:
-            raise APIException(
-                "Session filter token doesn't exists for current session")
-
-        def _create_qgs_expr(s, fidsin=None, fidsout=None):
-            """ Create qgs expression to save in db """
-
-            expr = f'$id IN ({fidsin})' if fidsin else f'$id NOT IN ({fidsout})'
-            return f'{s.qgs_expr} AND {expr}' if s else expr
-
-        if mode == 'create_update':
+            s, created = _get_sessiontokenfilter()
 
             l, created = s.stf_layers.get_or_create(
                 layer=self.layer,
@@ -438,10 +476,88 @@ class LayerVectorView(QGISLayerVectorViewMixin, BaseVectorApiView):
             token_data.update({
                 'filtertoken': s.token
             })
+
+        elif mode == 'save':
+
+            # Get qgs_expression from SessionTokenFilter
+            name = request_data.get('name')
+            if not name:
+                raise APIException(
+                    "'name' parameter is required for mode='save'.")
+
+            s, created = _get_sessiontokenfilter()
+
+            qgs_expr = s.stf_layers.get(layer=self.layer).qgs_expr
+            l, created = FilterLayerSaved.objects.update_or_create(
+                layer = self.layer,
+                user = request.user,
+                name = name,
+                defaults={'qgs_expr': qgs_expr}
+            )
+
+            token_data.update({
+                'layer': self.layer.qgs_layer_id,
+                'qgs_expression': qgs_expr,
+                'name': l.name,
+                'fid': l.pk,
+                'state': 'created' if created else 'updated'
+            })
+
+        elif mode == 'apply':
+
+            kwargs = _check_url_params_name_fid('apply')
+
+            fls = FilterLayerSaved.objects.get(
+                layer = self.layer,
+                user = request.user,
+                **kwargs
+            )
+
+            s, created = _get_sessiontokenfilter()
+
+            # Check if a recordo for surrent layer is present for update or create it
+            s.stf_layers.update_or_create(
+                layer=fls.layer,
+                defaults={
+                    'qgs_expr': fls.qgs_expr
+                }
+            )
+
+            # Return newone of current filter token
+            token_data.update({
+                'filtertoken': s.token
+            })
+
+
+        elif mode == 'delete_saved':
+
+            kwargs = _check_url_params_name_fid('delete_saved')
+
+            FilterLayerSaved.objects.get(
+                layer = self.layer,
+                user = request.user,
+                **kwargs
+            ).delete()
+
+        # mode='delete'
         else:
 
+            s, created = _get_sessiontokenfilter()
+
+            try:
+                l = s.stf_layers.get(layer=self.layer).delete()
+            except:
+                pass
+
+            token_data.update({
+                'filtertoken': s.token
+            })
+
             # delete session and
-            s.delete()
+            if len(s.stf_layers.all()) == 0:
+                token_data = {}
+                s.delete()
+
 
         self.results.update({'data': token_data})
